\name{cfaTools-package}
\alias{cfaTools-package}
\alias{cfaTools}
\docType{package}
\title{cfaTools: Tools for Large Scale Educational Assessment Analysis
}
\description{
Center for Assessment developed Tools for Large Scale Educational Assessment.
}

\details{
\tabular{ll}{
Package: \tab cfaTools\cr
Type: \tab Package\cr
Version: \tab 0.0-1.5\cr
Date: \tab 2021-8-13\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}

\author{Damian W. Betebenner \email{dbetebenner@nciea.org}, Adam Van Iwaarden \email{avaniwaarden@gmail.com}, Nathan Dadey \email{ndadey@nciea.org} and Joseph Martineau \email{jmartineau@nciea.org}}

\keyword{package}
